Move test helpers' tests to testing package
